BOT_Discord/requirements.txt
discord.py
python-dotenv   